# perf: Add multipacket async read target buffer caching