@roji 2019-11-06
> BDN isn't showing LOH allocations which is why the memory numbers look odd. Using profiler results you can see the savings.

As noted in https://github.com/dotnet/SqlClient/issues/245#issuecomment-550451589, we found out that it isn't about LOH allocations, but rather that allocations in other threads aren't taken into account. BDN 0.15.0 fixed this.