@David-Engel 2019-11-06
@Wraith2 Can you resolve the conflicts and merge the latest from master?