@Suchiman 2019-11-13
@roji BDN prior to that version uses `GetAllocatedBytesForCurrentThread` which also does not track LOH allocations correctly, see https://github.com/dotnet/coreclr/issues/27691#issuecomment-550305314