@roji 2019-11-13
@Suchiman thanks. I'm not sure if the discrepancies in this particular case were about the LOH allocations or about allocations in other threads, but it doesn't matter much.